Cập nhật

+++ src/main/java/com/example/gamestore/entity/Orders.java
@@ -10,10 +10,10 @@ public class Orders {
     private String order_id;
 	private String username; 
 	private Date order_date; 
-	private int total_amount; 
+	private double total_amount; 
 	private String status;
 	public Orders() {}
-	public Orders(String order_id, String username, Date order_date, int total_amount, String status) {
+	public Orders(String order_id, String username, Date order_date, double total_amount, String status) {
 		super();
 		this.order_id = order_id;
 		this.username = username;
@@ -39,10 +39,10 @@ public class Orders {
 	public void setOrder_date(Date order_date) {
 		this.order_date = order_date;
 	}
-	public int getTotal_amount() {
+	public double getTotal_amount() {
 		return total_amount;
 	}
-	public void setTotal_amount(int total_amount) {
+	public void setTotal_amount(double total_amount) {
 		this.total_amount = total_amount;
 	}
 	public String getStatus() {
